Tidying up public API, moving towards stabilisation
