--- .all-contributorsrc
@@ -1,151 +1,0 @@
-{
-  "projectName": "ward",
-  "projectOwner": "darrenburns",
-  "repoType": "github",
-  "repoHost": "https://github.com",
-  "files": [
-    "README.md"
-  ],
-  "imageSize": 60,
-  "commit": false,
-  "commitConvention": "none",
-  "contributors": [
-    {
-      "login": "darrenburns",
-      "name": "Darren Burns",
-      "avatar_url": "https://avatars0.githubusercontent.com/u/5740731?v=4",
-      "profile": "https://darrenburns.net",
-      "contributions": [
-        "code",
-        "doc",
-        "ideas",
-        "review",
-        "bug",
-        "example"
-      ]
-    },
-    {
-      "login": "khusrokarim",
-      "name": "khusrokarim",
-      "avatar_url": "https://avatars0.githubusercontent.com/u/1615476?v=4",
-      "profile": "https://github.com/khusrokarim",
-      "contributions": [
-        "ideas",
-        "code",
-        "bug"
-      ]
-    },
-    {
-      "login": "AlecJ",
-      "name": "Alec Jordan",
-      "avatar_url": "https://avatars2.githubusercontent.com/u/5054790?v=4",
-      "profile": "https://github.com/AlecJ",
-      "contributions": [
-        "code"
-      ]
-    },
-    {
-      "login": "CodeMouse92",
-      "name": "Jason C. McDonald",
-      "avatar_url": "https://avatars2.githubusercontent.com/u/7471402?v=4",
-      "profile": "https://www.indeliblebluepen.com",
-      "contributions": [
-        "code",
-        "ideas"
-      ]
-    },
-    {
-      "login": "AndydeCleyre",
-      "name": "Andy Kluger",
-      "avatar_url": "https://avatars3.githubusercontent.com/u/1787385?v=4",
-      "profile": "https://github.com/AndydeCleyre",
-      "contributions": [
-        "code",
-        "ideas"
-      ]
-    },
-    {
-      "login": "thebigmunch",
-      "name": "thebigmunch",
-      "avatar_url": "https://avatars0.githubusercontent.com/u/118418?v=4",
-      "profile": "https://forum.thebigmunch.me",
-      "contributions": [
-        "code"
-      ]
-    },
-    {
-      "login": "onlyanegg",
-      "name": "Tyler Couto",
-      "avatar_url": "https://avatars0.githubusercontent.com/u/7731128?v=4",
-      "profile": "https://github.com/onlyanegg",
-      "contributions": [
-        "code"
-      ]
-    },
-    {
-      "login": "thilp",
-      "name": "Thibaut Le Page",
-      "avatar_url": "https://avatars2.githubusercontent.com/u/968838?v=4",
-      "profile": "https://github.com/thilp",
-      "contributions": [
-        "code"
-      ]
-    },
-    {
-      "login": "DorianCzichotzki",
-      "name": "Dorian Czichotzki",
-      "avatar_url": "https://avatars1.githubusercontent.com/u/10177001?v=4",
-      "profile": "https://github.com/DorianCzichotzki",
-      "contributions": [
-        "code",
-        "ideas"
-      ]
-    },
-    {
-      "login": "jayeshathila",
-      "name": "jayesh hathila",
-      "avatar_url": "https://avatars0.githubusercontent.com/u/1469191?v=4",
-      "profile": "https://github.com/jayeshathila",
-      "contributions": [
-        "code"
-      ]
-    },
-    {
-      "login": "mandarvaze",
-      "name": "Mandar Vaze",
-      "avatar_url": "https://avatars1.githubusercontent.com/u/46438?v=4",
-      "profile": "https://mandarvaze.bitbucket.io/",
-      "contributions": [
-        "code"
-      ]
-    },
-    {
-      "login": "JoshKarpel",
-      "name": "Josh Karpel",
-      "avatar_url": "https://avatars2.githubusercontent.com/u/7133863?v=4",
-      "profile": "https://www.jtk.dev",
-      "contributions": [
-        "code"
-      ]
-    },
-    {
-      "login": "anlutro",
-      "name": "Andreas Lutro",
-      "avatar_url": "https://avatars0.githubusercontent.com/u/163093?v=4",
-      "profile": "https://www.lutro.me",
-      "contributions": [
-        "code"
-      ]
-    },
-    {
-      "login": "hoefling",
-      "name": "Oleg Höfling",
-      "avatar_url": "https://avatars1.githubusercontent.com/u/4455652?v=4",
-      "profile": "https://hoefling.io",
-      "contributions": [
-        "code"
-      ]
-    }
-  ],
-  "contributorsPerLine": 7
-}
