+++ tests/test_collect.py
@@ -9,7 +9,7 @@ import sys
 from cucumber_tag_expressions import parse
 
 from ward import fixture, raises, test
-from ward.collect import (
+from ward._collect import (
     _get_module_path,
     _handled_within,
     _is_excluded_module,
