+++ tests/test_debug.py
@@ -5,8 +5,8 @@ from types import SimpleNamespace
 from unittest import mock
 from unittest.mock import Mock
 
-from ward import test, debug
-from ward.debug import init_breakpointhooks, _breakpointhook, _get_debugger_hook, importlib
+from ward import test, _debug
+from ward._debug import init_breakpointhooks, _breakpointhook, _get_debugger_hook, importlib
 
 
 @test("init_breakpointhooks always patches pdb.set_trace")
@@ -18,21 +18,21 @@ def _():
 
 @test("init_breakpointhooks sets sys.breakpointhook when it's supported")
 def _():
-    old_func = debug._breakpoint_supported
-    debug._breakpoint_supported = lambda: True
+    old_func = _debug._breakpoint_supported
+    _debug._breakpoint_supported = lambda: True
     mock_sys = SimpleNamespace()
     init_breakpointhooks(pdb_module=Mock(), sys_module=mock_sys)
-    debug._breakpoint_supported = old_func
+    _debug._breakpoint_supported = old_func
     assert mock_sys.breakpointhook == _breakpointhook
 
 
 @test("init_breakpointhooks doesnt set breakpointhook when it's unsupported")
 def _():
-    old_func = debug._breakpoint_supported
-    debug._breakpoint_supported = lambda: False
+    old_func = _debug._breakpoint_supported
+    _debug._breakpoint_supported = lambda: False
     mock_sys = SimpleNamespace()
     init_breakpointhooks(pdb_module=Mock(), sys_module=mock_sys)
-    debug._breakpoint_supported = old_func
+    _debug._breakpoint_supported = old_func
     assert not hasattr(mock_sys, "breakpointhook")
