+++ tests/test_util.py
@@ -4,7 +4,7 @@ from pathlib import Path
 from ward import test, using, fixture
 from ward.testing import each
 from tests.utilities import make_project
-from ward.util import (
+from ward._utilities import (
     truncate,
     find_project_root,
     group_by,
