+++ tests/utilities.py
@@ -5,7 +5,8 @@ from pathlib import Path
 
 from ward import test, fixture
 from ward.fixtures import Fixture
-from ward.testing import Test, is_test_module_name
+from ward.testing import Test
+from ward._testing import is_test_module_name
 
 NUMBER_OF_TESTS = 5
 FORCE_TEST_PATH = Path("path/of/test").absolute()
