+++ ward/__init__.py
@@ -1,4 +1,4 @@
-"""A modern Python 3 test framework for finding and fixing flaws faster."""
+"""A modern Python 3 test framework with a focus on productivity and readability."""
 from ._ward_version import __version__
 from .expect import raises
 from .fixtures import fixture, using
