+++ ward/__main__.py
@@ -1,4 +1,4 @@
-from ward.run import run
+from ward._run import run
 
 if __name__ == "__main__":
     run()
