+++ ward/testing.py
@@ -3,44 +3,41 @@ import collections
 import functools
 import inspect
 import traceback
-import uuid
 from bdb import BdbQuit
-from collections import defaultdict
 from contextlib import ExitStack, closing, redirect_stderr, redirect_stdout
 from dataclasses import dataclass, field
 from enum import Enum, auto
 from io import StringIO
 from pathlib import Path
-from timeit import default_timer
-from types import MappingProxyType
 from typing import (
     Any,
     Callable,
     Dict,
     List,
     Optional,
-    Tuple,
     Union,
-    Iterable,
     Mapping,
+    Iterable,
     Collection,
 )
 
-from ward.errors import FixtureError, ParameterisationError
-from ward.fixtures import Fixture, FixtureCache, ScopeKey, is_fixture
+from ward._errors import FixtureError, ParameterisationError
+from ward._testing import (
+    Each,
+    _generate_id,
+    _FormatDict,
+    ParamMeta,
+    is_test_module_name,
+    COLLECTED_TESTS,
+    _Timer,
+)
+from ward._fixtures import ScopeKey, FixtureCache, is_fixture
+from ward.fixtures import Fixture
 from ward.models import Marker, Scope, SkipMarker, WardMeta, XfailMarker
-from ward.util import get_absolute_path
+from ward._utilities import get_absolute_path
 
 
-@dataclass
-class Each:
-    args: Tuple[Any]
-
-    def __getitem__(self, args):
-        return self.args[args]
-
-    def __len__(self):
-        return len(self.args)
+__all__ = ["each", "skip", "xfail", "Test", "test", "TestOutcome", "TestResult"]
 
 
 def each(*args):
@@ -99,21 +96,6 @@ def xfail(
     return wrapper
 
 
-def generate_id():
-    return uuid.uuid4().hex
-
-
-class FormatDict(dict):
-    def __missing__(self, key):
-        return "{" + key + "}"
-
-
-@dataclass
-class ParamMeta:
-    instance_index: int = 0
-    group_size: int = 1
-
-
 @dataclass
 class Test:
     """
@@ -122,7 +104,7 @@ class Test:
 
     fn: Callable
     module_name: str
-    id: str = field(default_factory=generate_id)
+    id: str = field(default_factory=_generate_id)
     marker: Optional[Marker] = None
     description: Optional[str] = None
     param_meta: Optional[ParamMeta] = field(default_factory=ParamMeta)
@@ -130,7 +112,7 @@ class Test:
     sout: StringIO = field(default_factory=StringIO)
     serr: StringIO = field(default_factory=StringIO)
     ward_meta: WardMeta = field(default_factory=WardMeta)
-    timer: Optional["Timer"] = None
+    timer: Optional["_Timer"] = None
     tags: List[str] = field(default_factory=list)
 
     def __hash__(self):
@@ -141,7 +123,7 @@ class Test:
 
     def run(self, cache: FixtureCache, dry_run=False) -> "TestResult":
         with ExitStack() as stack:
-            self.timer = stack.enter_context(Timer())
+            self.timer = stack.enter_context(_Timer())
             if self.capture_output:
                 stack.enter_context(redirect_stdout(self.sout))
                 stack.enter_context(redirect_stderr(self.serr))
@@ -272,7 +254,7 @@ class Test:
 
         generated_tests = []
         for instance_index in range(number_of_instances):
-            test = Test(
+            generated_test = Test(
                 fn=self.fn,
                 module_name=self.module_name,
                 marker=self.marker,
@@ -282,7 +264,7 @@ class Test:
                 ),
                 capture_output=self.capture_output,
             )
-            generated_tests.append(test)
+            generated_tests.append(generated_test)
         return generated_tests
 
     def find_number_of_instances(self) -> int:
@@ -321,7 +303,7 @@ class Test:
         Returns the newly updated description.
         """
 
-        format_dict = FormatDict(**args)
+        format_dict = _FormatDict(**args)
         if not self.description:
             self.description = ""
 
@@ -333,9 +315,102 @@ class Test:
         return self.description
 
 
+def test(description: str, *args, tags=None, **kwargs):
+    def decorator_test(func):
+        unwrapped = inspect.unwrap(func)
+        module_name: str = unwrapped.__module__
+        is_home_module: bool = "." not in module_name
+        if is_test_module_name(module_name) and is_home_module:
+            force_path: Path = kwargs.get("_force_path")
+            if force_path:
+                path = force_path.absolute()
+            else:
+                path = get_absolute_path(unwrapped)
+
+            if hasattr(unwrapped, "ward_meta"):
+                unwrapped.ward_meta.description = description
+                unwrapped.ward_meta.tags = tags
+                unwrapped.ward_meta.path = path
+            else:
+                unwrapped.ward_meta = WardMeta(
+                    description=description, tags=tags, path=path,
+                )
+
+            collect_into = kwargs.get("_collect_into", COLLECTED_TESTS)
+            collect_into[path].append(unwrapped)
+
+            @functools.wraps(func)
+            def wrapper(*args, **kwargs):
+                return func(*args, **kwargs)
+
+            return wrapper
+
+        return func
+
+    return decorator_test
+
+
+class TestOutcome(Enum):
+    PASS = auto()
+    FAIL = auto()
+    SKIP = auto()
+    XFAIL = auto()  # expected fail
+    XPASS = auto()  # unexpected pass
+    DRYRUN = auto()  # tests arent executed during dryruns
+
+    @property
+    def display_char(self):
+        display_chars = {
+            TestOutcome.PASS: ".",
+            TestOutcome.FAIL: "F",
+            TestOutcome.SKIP: "-",
+            TestOutcome.XPASS: "U",
+            TestOutcome.XFAIL: "x",
+            TestOutcome.DRYRUN: ".",
+        }
+        assert len(display_chars) == len(TestOutcome)
+        return display_chars[self]
+
+    @property
+    def display_name(self):
+        display_names = {
+            TestOutcome.PASS: "Passes",
+            TestOutcome.FAIL: "Failures",
+            TestOutcome.SKIP: "Skips",
+            TestOutcome.XPASS: "Unexpected Passes",
+            TestOutcome.XFAIL: "Expected Failures",
+            TestOutcome.DRYRUN: "Dry-runs",
+        }
+        assert len(display_names) == len(TestOutcome)
+        return display_names[self]
+
+
+@dataclass
+class TestResult:
+    test: Test
+    outcome: TestOutcome
+    error: Optional[Exception] = None
+    message: str = ""
+    captured_stdout: str = ""
+    captured_stderr: str = ""
+
+
+def fixtures_used_directly_by_tests(
+    tests: Iterable["Test"],
+) -> Mapping[Fixture, Collection["Test"]]:
+    test_to_fixtures = {t: t.resolver.fixtures for t in tests}
+
+    fixture_to_tests = collections.defaultdict(list)
+    for test, used_fixtures in test_to_fixtures.items():
+        for fix in used_fixtures.values():
+            fixture_to_tests[fix].append(test)
+
+    return fixture_to_tests
+
+
 @dataclass
 class TestArgumentResolver:
-    test: Test
+    test: "Test"
     iteration: int
 
     def resolve_args(self, cache: FixtureCache) -> Dict[str, Any]:
@@ -483,7 +558,8 @@ class TestArgumentResolver:
         cache.cache_fixture(fixture, scope_key)
         return fixture
 
-    def _unpack_resolved(self, fixture_dict: Dict[str, Any]) -> Dict[str, Any]:
+    @staticmethod
+    def _unpack_resolved(fixture_dict: Dict[str, Any]) -> Dict[str, Any]:
         resolved_vals = {}
         for (k, arg) in fixture_dict.items():
             if isinstance(arg, Fixture):
@@ -491,122 +567,3 @@ class TestArgumentResolver:
             else:
                 resolved_vals[k] = arg
         return resolved_vals
-
-
-def fixtures_used_directly_by_tests(
-    tests: Iterable[Test],
-) -> Mapping[Fixture, Collection[Test]]:
-    test_to_fixtures = {test: test.resolver.fixtures for test in tests}
-
-    fixture_to_tests = collections.defaultdict(list)
-    for test, used_fixtures in test_to_fixtures.items():
-        for fix in used_fixtures.values():
-            fixture_to_tests[fix].append(test)
-
-    return fixture_to_tests
-
-
-def is_test_module_name(module_name: str) -> bool:
-    return module_name.startswith("test_") or module_name.endswith("_test")
-
-
-# Tests declared with the name _, and with the @test decorator
-# have to be stored in here, so that they can later be retrieved.
-# They cannot be retrieved directly from the module due to name
-# clashes. When we're later looking for tests inside the module,
-# we can retrieve any anonymous tests from this dict.
-# Map of module absolute Path to list of tests in the module
-anonymous_tests: Dict[Path, List[Callable]] = defaultdict(list)
-
-
-def test(description: str, *args, tags=None, **kwargs):
-    def decorator_test(func):
-        unwrapped = inspect.unwrap(func)
-        module_name: str = unwrapped.__module__
-        is_home_module: bool = "." not in module_name
-        if is_test_module_name(module_name) and is_home_module:
-            force_path: Path = kwargs.get("_force_path")
-            if force_path:
-                path = force_path.absolute()
-            else:
-                path = get_absolute_path(unwrapped)
-
-            if hasattr(unwrapped, "ward_meta"):
-                unwrapped.ward_meta.description = description
-                unwrapped.ward_meta.tags = tags
-                unwrapped.ward_meta.path = path
-            else:
-                unwrapped.ward_meta = WardMeta(
-                    description=description, tags=tags, path=path,
-                )
-
-            collect_into = kwargs.get("_collect_into", anonymous_tests)
-            collect_into[path].append(unwrapped)
-
-            @functools.wraps(func)
-            def wrapper(*args, **kwargs):
-                return func(*args, **kwargs)
-
-            return wrapper
-
-        return func
-
-    return decorator_test
-
-
-class TestOutcome(Enum):
-    PASS = auto()
-    FAIL = auto()
-    SKIP = auto()
-    XFAIL = auto()  # expected fail
-    XPASS = auto()  # unexpected pass
-    DRYRUN = auto()  # tests arent executed during dryruns
-
-    @property
-    def display_char(self):
-        display_chars = {
-            TestOutcome.PASS: ".",
-            TestOutcome.FAIL: "F",
-            TestOutcome.SKIP: "-",
-            TestOutcome.XPASS: "U",
-            TestOutcome.XFAIL: "x",
-            TestOutcome.DRYRUN: ".",
-        }
-        assert len(display_chars) == len(TestOutcome)
-        return display_chars[self]
-
-    @property
-    def display_name(self):
-        display_names = {
-            TestOutcome.PASS: "Passes",
-            TestOutcome.FAIL: "Failures",
-            TestOutcome.SKIP: "Skips",
-            TestOutcome.XPASS: "Unexpected Passes",
-            TestOutcome.XFAIL: "Expected Failures",
-            TestOutcome.DRYRUN: "Dry-runs",
-        }
-        assert len(display_names) == len(TestOutcome)
-        return display_names[self]
-
-
-@dataclass
-class TestResult:
-    test: Test
-    outcome: TestOutcome
-    error: Optional[Exception] = None
-    message: str = ""
-    captured_stdout: str = ""
-    captured_stderr: str = ""
-
-
-class Timer:
-    def __init__(self):
-        self._start_time = None
-        self.duration = None
-
-    def __enter__(self):
-        self._start_time = default_timer()
-        return self
-
-    def __exit__(self, *args):
-        self.duration = default_timer() - self._start_time
